requirement.txt
ultralytics==8.1.0
opencv-python>=4.8.0
pyserial>=3.5
roboflow>=1.1.0
networkx>=3.1
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0